[DF-9163] c-style casts fixed/reverted
Former-commit-id: 06ed0d8a0b88f9e8e37d096b273fac948a1df394

+++ Tools/QuickEd/Classes/UI/Preview/Ruler/RulerWidget.cpp
@@ -119,7 +119,7 @@ void RulerWidget::DrawScale(QPainter& painter, int tickStep, int tickStartPos, i
         return;
     }
 
-    int numberOffset = static_cast<int>(std::ceil(static_cast<float>(settings.startPos) / static_cast<float>(tickStep)) * static_cast<float>(tickStep));
+    int numberOffset = std::ceilf((float)(settings.startPos) / (float)(tickStep)) * (float)tickStep;
     int tickOffset = numberOffset - settings.startPos;
 
     int fontPos = 0;
